requirements/requirements.txt
click==7.1.2
click-completion==0.5.2
Jinja2>=2.11
dictdiffer==0.8.1
google-cloud-logging==1.15.1
google-cloud-bigquery>=1.26.0
google-api-python-client>=1.12.1
